Adicionado tratamento de exceptions, modificacao no voto, remodelagem de tabelas com flyway

+++ backend/src/main/java/com/imdb/filmes/repository/VotoRepository.java
@@ -3,6 +3,6 @@ package com.imdb.filmes.repository;
 import com.imdb.filmes.model.Voto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface VotoRepository extends JpaRepository<Voto, Integer> {
+public interface VotoRepository extends JpaRepository<Voto, Long> {
 
 }
